+++ Engine/AudioHelper.hpp
@@ -21,6 +21,6 @@ public:
     static void ChangeSamplePosition(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance, float position);
     static void ChangeSampleSpeed(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance, float speed);
     static unsigned int GetSampleLength(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance);
-    static unsigned int GetSamplePosition(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance);
+    static double GetSamplePosition(std::shared_ptr<ALLEGRO_SAMPLE_INSTANCE> sample_instance);
 };
 #endif   // AUDIOHELPER_HPP
